// src/app/core/models/service/service.ts
export class ServiceModel {
    _id: string;
    title: string;
    category: string
    duration: number;
    charges: number;
    commission: number;
    staffcommission : any[];
    availability: object;
    staffavailability: any[];
    breaktime: any[];
    staff: any[];
    gallery : any[];
    type: string;
    property: object;
    branchid: string;
    status: string;
}